refactor(session): unify abort registry into deep coordinator and collapse seams
