+++ packages/server/src/llm/stream/session-run-coordinator.ts
@@ -1,5 +1,9 @@
 import type { PrefixedString } from '@stitch/shared/id';
 
+import * as Log from '@/lib/log.js';
+
+const log = Log.create({ service: 'abort-registry' });
+
 type SessionRun = (abortSignal: AbortSignal) => Promise<void>;
 
 type QueuedRun = { run: SessionRun; resolve: () => void; reject: (error: unknown) => void };
@@ -58,4 +62,43 @@ export function cancelQueuedSessionRuns(sessionId: PrefixedString<'ses'>): void 
 
 export function isSessionRunActive(sessionId: PrefixedString<'ses'>): boolean {
   return Boolean(queues.get(sessionId)?.activeController);
+}
+
+// ---------------------------------------------------------------------------
+// Generic abort registry (collapsed from lib/abort-registry.ts)
+// Private per-session AbortControllers for background/child sessions.
+// ---------------------------------------------------------------------------
+
+const genericRegistry = new Map<PrefixedString<'ses'>, AbortController>();
+
+export function register(sessionId: PrefixedString<'ses'>): AbortSignal {
+  const existing = genericRegistry.get(sessionId);
+  if (existing) {
+    log.warn(
+      { event: 'stream.abort.registry_reregister', sessionId },
+      'aborting existing controller before re-registering',
+    );
+    existing.abort();
+  }
+
+  const controller = new AbortController();
+  genericRegistry.set(sessionId, controller);
+  return controller.signal;
+}
+
+export function abort(sessionId: PrefixedString<'ses'>): void {
+  const controller = genericRegistry.get(sessionId);
+  if (!controller) return;
+  log.info({ event: 'stream.abort.registry_abort', sessionId }, 'aborting session');
+  controller.abort();
+  genericRegistry.delete(sessionId);
+}
+
+export function cleanup(sessionId: PrefixedString<'ses'>): void {
+  genericRegistry.delete(sessionId);
+}
+
+export function abortSession(sessionId: PrefixedString<'ses'>): void {
+  abortActiveSessionRun(sessionId);
+  cancelQueuedSessionRuns(sessionId);
 }
